Added a new mascot.
The mascot is derived from the work of the wonderful Lauren Dawson. She
has given permission for its use as long as it is credited. See pasted
emails below (Address removed for privacy concerns):

"""
Lauren Dawson <***********@gmail.com>
6 Sep (1 day ago)

to james
HEY james im so sorry about the late response ! im definitely alright with you using one of the gryphons as long as the credits there, thanks for asking - im gonna take a guess at the one youre thinking of. . ...  .. .. is it . . .. . . . . . . . . .. .  . ... .  .. .  the vulture/sphinx ?

On Tue, Aug 30, 2016 at 5:34 AM, Storenvy Member - James Saunders <noreply@storenvy.com> wrote:
Hi,

I run a small opensource software project (among other things) called Gryphon. It does something pretty technical and difficult to explain, but it is a little tradition in my company that we create a mascot for each of our projects, of the animal shooting lasers out of its eyes. We have a current mascot but it is not very good and we came across your alternative Gryphon images. We were wondering if we could use a version of one of them as the base for our mascot. If you email me I will send you the image we want to use. I fully understand if you are hesitant, I know that you make a living from this, but you will be fully credited and we are not making money form sharing this project. Please let me know either way, you can check out the project at https://github.com/BusinessOptics/gryphon .

Cheers,

James.

Name: James Saunders
Email: james@businessoptics.biz
Phone: +27797709290
"""